Fixed issue upon chatroom page reload
When user is logged in, the chatroom page was throwing an error
regarding component rendering, thus the issue is fixed.

An issue persists however regarding the loading state of protected
routes/pages upon reload (in particular, you can see a glimpse of the
login page upon reloading the chatroom page when logged in). Will fix
in a later commit.